requirements.txt
python-telegram-bot==21.6
pytest==8.3.3
